+++ packages/backend/src/nest/tor/tor.service.ts
@@ -227,28 +227,6 @@ export class Tor extends EventEmitter implements OnModuleInit {
 
       this.process.stdout.on('data', (data: any) => {
         this.logger(data.toString())
-        const info = data.toString()
-
-        const textIndex = info.indexOf('):') + 2
-        const text = info.slice(textIndex).trim()
-
-        switch (text) {
-          case ConnectionProcessInfo.TOR_1:
-            this.emit(SocketActionTypes.CONNECTION_PROCESS_INFO, ConnectionProcessInfo.TOR_1)
-            break
-          case ConnectionProcessInfo.TOR_2:
-            this.emit(SocketActionTypes.CONNECTION_PROCESS_INFO, ConnectionProcessInfo.TOR_2)
-            break
-          case ConnectionProcessInfo.TOR_3:
-            this.emit(SocketActionTypes.CONNECTION_PROCESS_INFO, ConnectionProcessInfo.TOR_3)
-            break
-          case ConnectionProcessInfo.TOR_4:
-            this.emit(SocketActionTypes.CONNECTION_PROCESS_INFO, ConnectionProcessInfo.TOR_4)
-            break
-          case ConnectionProcessInfo.TOR_5:
-            this.emit(SocketActionTypes.CONNECTION_PROCESS_INFO, ConnectionProcessInfo.TOR_5)
-            break
-        }
 
         const regexp = /Bootstrapped 0/
         if (regexp.test(data.toString())) {