+++ packages/desktop/src/renderer/components/LoadingPanel/JoiningPanelComponent.test.tsx
@@ -71,7 +71,7 @@ describe('Create JoiningPanelComponent', () => {
                 style="width: 600px;"
               >
                 <div
-                  class="MuiGrid-root MuiGrid-container JoiningPanelComponentroot css-kpkqnh-MuiGrid-root"
+                  class="MuiGrid-root MuiGrid-container JoiningPanelComponentroot css-1vn3v3s-MuiGrid-root"
                 >
                   <div
                     class="MuiGrid-root MuiGrid-container MuiGrid-direction-xs-column JoiningPanelComponentcontentWrapper css-t0zib5-MuiGrid-root"
@@ -95,11 +95,14 @@ describe('Create JoiningPanelComponent', () => {
                         <div
                           class="JoiningPanelComponentprogress"
                         />
+                        <div
+                          class=""
+                        />
                       </div>
                       <p
                         class="MuiTypography-root MuiTypography-body2 css-16d47hw-MuiTypography-root"
                       >
-                        Initializing storage
+                        Initialized backend modules
                       </p>
                     </div>
                     <p
@@ -205,7 +208,7 @@ describe('Create JoiningPanelComponent', () => {
                 style="width: 600px;"
               >
                 <div
-                  class="MuiGrid-root MuiGrid-container JoiningPanelComponentroot css-kpkqnh-MuiGrid-root"
+                  class="MuiGrid-root MuiGrid-container JoiningPanelComponentroot css-1vn3v3s-MuiGrid-root"
                 >
                   <div
                     class="MuiGrid-root MuiGrid-container MuiGrid-direction-xs-column JoiningPanelComponentcontentWrapper css-t0zib5-MuiGrid-root"
@@ -229,11 +232,14 @@ describe('Create JoiningPanelComponent', () => {
                         <div
                           class="JoiningPanelComponentprogress"
                         />
+                        <div
+                          class=""
+                        />
                       </div>
                       <p
                         class="MuiTypography-root MuiTypography-body2 css-16d47hw-MuiTypography-root"
                       >
-                        Initializing storage
+                        Initialized backend modules
                       </p>
                     </div>
                     <p