+++ packages/mobile/src/components/ConnectionProcess/ConnectionProcess.types.ts
@@ -1,4 +1,6 @@
+import { ConnectionProcessInfo } from '@quiet/types'
+
 export interface ConnectionProcessComponentProps {
-  connectionProcess: { number: number; text: string }
+  connectionProcess: { number: number; text: ConnectionProcessInfo }
   openUrl: (url: string) => void
 }