+++ packages/mobile/src/tests/joining.process.test.tsx
@@ -54,12 +54,12 @@ describe('Joining process', () => {
     const processText = screen.getByTestId('connection-process-text')
     expect(processText.props.children).toEqual('Connecting process started')
 
-    store.dispatch(connection.actions.setConnectionProcess(ConnectionProcessInfo.INITIALIZING_LIBP2P))
+    store.dispatch(connection.actions.setConnectionProcess(ConnectionProcessInfo.INITIALIZING_IPFS))
     await act(async () => {})
 
     const processText2 = screen.getByTestId('connection-process-text')
     console.log(processText2.props)
-    expect(processText2.props.children).toEqual('Initializing libp2p')
+    expect(processText2.props.children).toEqual('Initialized backend modules')
 
     store.dispatch(connection.actions.setConnectionProcess(ConnectionProcessInfo.LAUNCHED_COMMUNITY))
     await act(async () => {})