+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+[unreleased]
+
+* Better descriptions of the joining process
+
 [2.0.3-alpha.5]
 
 * Fix network data proceeding when using custom protocol multiple times #1847